Refactored GUI
Refactored GUI to make it visually cleaner. Moved waves below rustnes, so that rustnes can use them (Let me know if I need to move it back). Added a navigation bar with a file context menu, and a control bar with play, pause, and volume. Left the middle of the window empty for now.